move export logic to back class
